pyproject.toml
[tool.poetry]
name = "todoer"
version = "0.1.0"
description = ""
authors = ["Andrew Veselov <vesel41223@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
django = "^4.2.5"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.0"
django-extensions = "^3.2.3"
ptpython = "^3.0.23"
marshmallow = "^3.20.1"
marshmallow-dataclass = "^8.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
